refactor(types): move the types to config
