+++ src/config/Pokemons/Pokemons.types.ts
