+++ src/config/Store/Store.ts
@@ -4,9 +4,8 @@ import thunkMiddleware from "redux-thunk";
 import { rootReducer } from "@/reducers/Root";
 import { isDev } from "@/utilities/Env";
 
-const devTools =
-    typeof window !== undefined &&
-    (window as any).__REDUX_DEVTOOLS_EXTENSION_COMPOSE__;
+const devToolsName = "__REDUX_DEVTOOLS_EXTENSION_COMPOSE__";
+const devTools = typeof window !== undefined && (window as any)[devToolsName];
 
 const enchancer = isDev ? devTools || compose : compose;
 const middleware = applyMiddleware(thunkMiddleware);
