+++ src/utilities/HTTP/HTTP.types.ts
@@ -1,6 +1,6 @@
 import { AxiosResponse } from "axios";
 
-import { Pokemon } from "@/types/pokemons";
+import { Pokemon } from "@/config/Pokemons/Pokemons.types";
 
 export type PokemonsResponse = Promise<AxiosResponse<Pokemon[]>>;
 export type UndefinedResponse = Promise<AxiosResponse<undefined>>;
